Moved Popup to new window
