+++ src/contentScripts/index.ts
@@ -1,5 +1,3 @@
 import tabBlocker from "./tabBlocker";
 
-import passwordEncryption from "./passwordEncryption";
-
-export { tabBlocker, passwordEncryption };
+export { tabBlocker };
